convert queue tests to mocha
